Task4

+++ Tyuiu.ZhuriloNA.Sprint0.Task4.V0.Lib/DataService.cs
@@ -20,7 +20,7 @@ namespace Tyuiu.ZhuriloNA.Sprint0.Task4.V0.Lib
         {
             return (a * b);
         }
-        public static int Division(int a, int b)
+        public static double Division(int a, int b)
         {
             return (a / b);
         }
